+++ flashback/iterating/functions.py
@@ -113,3 +113,26 @@ def partition(predicate, iterable):
             falses.append(item)
 
     return (tuple(trues), tuple(falses))
+
+
+def uniq(iterable):
+    """
+    Removes duplicates items from `iterable` while respecting their apparition order.
+
+    Params:
+        - `iterable (Iterable<Any>)` the iterable to remove duplicates from
+
+    Returns:
+        - `tuple<Any>` the iterable without duplicates
+    """
+    uniq = []
+    seen = set()
+
+    for item in iterable:
+        if item in seen:
+            continue
+
+        uniq.append(item)
+        seen.add(item)
+
+    return tuple(uniq)